media: Fix widevine_cdm_manifest target
On Linux/ChromeOS, we don't use manifest.json for Widevine CDM.
Therefore the "widevine_cdm_manifest_file" list is empty. Then, in a GYP
build, copying an empty list to a destination will actually cause the
whole folder where the gyp file is located to be copied. See BUG for
more details.

This CL fixes this issue by only coping the manifest when the list is
not empty.

BUG=615500
TEST=Tested on Linux where the list is empty and on Windows where the
list is not empty.

Review-Url: https://codereview.chromium.org/2023383003
Cr-Original-Commit-Position: refs/heads/master@{#397308}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 35e11eb20c09bef5c5b8ba0eb3b352ab314d94d3