PrePaintTreeWalk should walk multicol spanner through the placeholder
The LayoutObject of a multicol spanner is placed in the layout tree
according to the DOM order, but it actually belongs to the multicol
container, not its DOM hierarchy. During PrePaintTreeWalk we should
follow the link from the placeholder to the spanner, instead of in
layout tree order.

Review-Url: https://codereview.chromium.org/2031923002
Cr-Original-Commit-Position: refs/heads/master@{#397463}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b3ea039dd70cb06308ecb4087e7a1f9d90005c2a