Enable pylint warnings on presubmit checks.
In http://crbug.com/493941, it was mentioned that we could fix all current pylint warnings before enabling pylint warnings in the presubmit -- but actually we could also enable warnings in the presubmit now without causing any problems; pylint will only print warnings about lines that are changed in a CL, and turning on warnings now could be potentially helpful in cleaning up the code.

And turning on pylint warnings in check-webkit-style and the presubmit is just a one line change; any reason not to just do it now?

NOPRESUBMIT=true
BUG=493941

Review-Url: https://codereview.chromium.org/2010173003
Cr-Original-Commit-Position: refs/heads/master@{#397780}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d198da47ae0b4952bdfb3f15640bf1071cd86e3f