SkFontConfigInterface createTypeface to makeTypeface.
The createTypeface method (which returned a bare pointer to a reference
counted object) is being replaced with makeTypeface (which returns a
smart pointer to a reference counted object). This removes some manual
reference counting in Chromium and allows for removing the deprecated
createTypeface method.

Review-Url: https://codereview.chromium.org/2037133002
Cr-Original-Commit-Position: refs/heads/master@{#397794}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d8ada9acb3bb8287191e193a2c1c9f85b2d124a