Implement StringImpl sharing for StringView::toString().
Prior to https://codereview.chromium.org/2007103003 StringBuilder had an
optimization that would share the StringImpl when calling append() with
a String the first time. This meant that doing element.textContent on
an Element with a single Text node child wouldn't need to string copy.

The optimization also helps the parser, removing a string copy when
flushing text in the HTML parser (flushPendingText) when there's a
single text that was appended, and the text is a reasonable length
(<= 65536 bytes).

This optimization further helps any code that would potentially append
an AtomicString to a StringBuilder and then do toAtomicString()
or toString() by ensuring we return the same AtomicString again
avoiding a bunch of work, and allowing the atomic string optimization
to propagate to the result.

This patch restores the StringBuilder optimization by adding a
StringImpl* member to StringView and allowing toString() and
toAtomicString() to return that member when the offset is zero and the
length is the same as the input string.

This patch also collapses the m_is8Bit flag into the m_impl by just
reusing the field to know if the StringView is 8bit. We now store the
shared empty 8bit or 16bit strings in the impl field for views created
from sources that are not backed by StringImpl. Doing this keeps the
StringView class to be only 2 ptrs and an unsigned wide.

I also further simplified the code by using the union members directly
and by adding a void* bytes() getter to StringImpl which removed the
branch on is8Bit for the common case of constructing a StringView from
a String or StringImpl with no offset, and the full length.

BUG=615174

Review-Url: https://codereview.chromium.org/2033293002
Cr-Original-Commit-Position: refs/heads/master@{#397986}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 580cbf5d215fee758dd5c447cf03885b07bc9d2e