Make CSSComputedStyledeclaration::getPropertyCSSValue return const
Make CSSComputedStyledeclaration::getPropertyCSSValue() return a const
CSSValue* instead of a regular CSSValue*. This required changing
CSSStyleDeclaration::getPropertyCSSValueInternal() return a const
CSSValue* as well, and marking CSSValueList::copy() as a const method
(which it already was, it was just missing the const modifier). This is
pre-work for making CSSValueList store const CSSValues.

This patch is mostly mechanical changes, but does contain one logic
change: the static method mergeTextDecorationValues in EditingStyle
needs to be changed since it modifies a CSSValueList that points inside
the mutable style. Instead, the function is changed to take two
CSSValueLists and return a new one, and the callsite is changed to
replace the old CSSValueList with the new one in the style rather than
modifying the old CSSValueList in-place.

BUG=526586

Review-Url: https://codereview.chromium.org/2034013002
Cr-Original-Commit-Position: refs/heads/master@{#398151}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5896ec297d9240b691e7f70a239498b795531dc1