Fixes ASan crash for an embedded Blink-in-JS component.
In the test case (in this patch) it appears the Blink-in-JS component tries to run JS during document creation.
However there is a ScriptForbidden scope which throws a "Uncaught Error: Script execution is forbidden." (probably because it is being created in this weird place?)

This patch re-adds the simple check that was removed in https://codereview.chromium.org/1885833002 which checked if the toDOMWindow(isolate->GetEnteredContext()) was null.
(now the check is just isolate->GetEnteredContext()->IsEmpty()).

BUG=617104

Review-Url: https://codereview.chromium.org/2039333002
Cr-Original-Commit-Position: refs/heads/master@{#398310}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2deeee5b5604eb997fa053a1b6ab4268c662596f