Change the timer to record the actual drawing time
Currently in the BaseRenderingContext2D.cpp, we put the scoped timer at
the very beginning of some API functions. But that includes timing for
some early exits cases. This CL makes changes such that we do not record
the time for the early exits cases

BUG=612585

Review-Url: https://codereview.chromium.org/2047643004
Cr-Original-Commit-Position: refs/heads/master@{#398381}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38a83802d7b2af4fc25bbb6d20490fda1cba7700