Add RELEASE_ASSERT(!needsLayout()) in FrameView::layout()
Now we know that the failure of RELEASE_ASSERT(!needslayout())
in FrameView is because that FrameView::layout() returns with
dirty layout in some cases. Add RELEASE_ASSERT(!needslayout())
in FrameView::layout() to find the detail reasons.

BTW, let adjustViewSizeAndLayout() return a bool indicating if
a layout is actually done to avoid duplicated work after the
function in FrameView::layout().

BUG=590856

Review-Url: https://codereview.chromium.org/2041983004
Cr-Original-Commit-Position: refs/heads/master@{#398436}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 46d2c68225a970a166bd6eeba4d4cee56ae09cfa