Add 'selected' boolean to 'ShippingOption'
ShippingOption.selected is a new boolean member variable used in
PaymentRequest.

"This is set to true to indicate that this is the default selected
ShippingOption in a sequence. User agents should display this option by
default in the user interface."

The website now has a more powerful way to specify the selected shipping
option. Prior to this patch, specifying a single shipping option would
result in that option being selected as the default. After this patch,
only the option that has "selected: true" is selected by default. This
makes calculation of totals more deterministic, as the web developer now
knows which shipping option is selected in UI.

https://w3c.github.io/browser-payment-api/specs/paymentrequest.html#idl-def-shippingoption
https://github.com/w3c/browser-payment-api/commit/684603787c80d46d06b76f73b00a28a938e4b291

BUG=616948

Review-Url: https://codereview.chromium.org/2045833002
Cr-Original-Commit-Position: refs/heads/master@{#398471}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d784539e0ff790124c695b9e704752fa8e0ce525