[css-grid] Fix definite/indefinite size detection on block axis
LayoutBox::hasDefiniteLogicalHeight() was wrong as
it was not taking into account several cases
where the height is definite/indefinite.
E.g. it was not checking the special situations for flex items,
that despite of having height "auto" can definite sometimes.

As the method was buggy, this patch is directly removing it
and using LayoutBox::percentageLogicalHeightIsResolvable() instead.

Add new test to check that the flexbox case works as expected.

BUG=617876
TEST=fast/css-grid-layout/flex-item-grid-container-percentage-rows.html

Review-Url: https://codereview.chromium.org/2037383002
Cr-Original-Commit-Position: refs/heads/master@{#398680}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f8393437ca2c05d3334011444260adac7e82530c