Reorganize bindings test IDL files
Rename the various IDL files used to test bindings for partial interfaces.
The intent is to make it clearer which files are extending a particular
interface. The naming convention follows the suggestion in an earlier CL:
https://codereview.chromium.org/1884423002/#msg20

This prepares for a subsequent CL to fix the bindings for partial
interfaces for the [OriginTrialEnabled] extended attribute.

BUG=585656

Review-Url: https://codereview.chromium.org/2049843003
Cr-Original-Commit-Position: refs/heads/master@{#398856}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9a4335504c0d6238fe5a26832f916ebd16f9a0e