WebRTCCertificateGenerator dropping unused URL parameters.
Non-functional clean-up. The generator previously needed "url" and
"first party for cookies" because its implementation was using an identity store
that had access to storage for caching. The implementation has since been
replaced not to do this and the parameters are unused. So this CL removes them.

BUG=webrtc:5708, webrtc:5707

Review-Url: https://codereview.chromium.org/2030373002
Cr-Original-Commit-Position: refs/heads/master@{#398857}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3ae34f35b4618849690e871d76b0b1461ba08378