Notify the HTMLDocumentParser on document element available
This patch adds logic to the HTMLHtmlElement so that it can notify
the HTMLDocumentParser immediately when the document element is available.
This is triggered in HTMLHtmlElement::insertedByParser. Additionally,
logic to dispatch notifications to the embedder is  refactored into
insertedByParser.

insertedByParser initializes the ApplicationCache, then sends the
notification to the HTMLDocumentParser to issue any queued preloads (the
preloads wait for ApplicationCache so they don't download needlessly).
Next, the method dispatches notifications to embedders, via
FrameLoader::dispatchDocumentElementAvailable and
FrameLoader::runscriptsAtDocumentAvailable.

Note the latter is used by extensions, which can run extremely expensive
content scripts. This patch changes ordering so that preloads can be sent
before this happens.

BUG=618101

Review-Url: https://codereview.chromium.org/2042413002
Cr-Original-Commit-Position: refs/heads/master@{#398937}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d63bd15666f24c048b0caa641cace3eec9460769