cc: Put to_target and to_screen behind an accessor.
Target-related information will eventually be removed from transform
tree and computed when needed.

This CL wraps the target-related information in transform tree in an
accessor, so that we can hide the implementations of the accessor, thus
we can make the code work with either the old way or the new way of
computing the information each time we need it.

BUG=597721
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review-Url: https://codereview.chromium.org/2032213002
Cr-Original-Commit-Position: refs/heads/master@{#398989}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2c5f8d42b4bc2282953fb4c724831b8d7aac98f1