[Typed OM] Implement FilteredComputedStylePropertyMap
This is needed for the CSS Paint API, as inside the paint callback only
properties listed as inputProperties should exist in the style map.

FilteredComputedStylePropertyMap takes a list of both the native and
custom properties and accessors will throw if the property isn't in this
list.

BUG=578252

Review-Url: https://codereview.chromium.org/2054633002
Cr-Original-Commit-Position: refs/heads/master@{#398992}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4382d59bb2da192a3ee29f9ae454afe348c4056e