Clamp shown ratio when setting main thread top controls constraints.
We already do this on the compositor thread so this just applies matching
behaviour on the main thread. When we constrain the top controls to being shown
or hidden, we set the currently shown ratio to 1 or 0, respectively, at the time
the constraint is set, rather than waiting for a compositor commit to reflect
the delta back to the main thread.

BUG=476541

Review-Url: https://codereview.chromium.org/2044133002
Cr-Original-Commit-Position: refs/heads/master@{#399036}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bd10cc1ecbbcc95f143d19911a6d3b81b8043d65