Move WTF string to number converter functions to StringToNumber.h
This allows various things to call these methods without having to
depend on the WTFString.h header which was a strange place for
a bunch of functions that take LChar* and UChar*'s.

No functional changes are made, the code is just moved around.

BUG=615174

Review-Url: https://codereview.chromium.org/2058613002
Cr-Original-Commit-Position: refs/heads/master@{#399095}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7dc7c0fc073b84f3050726b52ceec1bbcb304ba3