[css-grid] Avoid unneeded calls to percentageLogicalHeightIsResolvable()
r398535 was introducing a performance regression on grid layout tests.
The problem is that it was calling LayoutGrid::hasDefiniteLogicalSize()
which was very slow.

r398680 was getting rid of hasDefiniteLogicalSize() and replacing it
by calls to LayoutBox::percentageLogicalHeightIsResolvable().
Despite being faster it's still an expensive operation and
we don't need to do it always.

So in this patch we're avoid calling percentageLogicalHeightIsResolvable()
when the min and max track breadths are not percentages.

BUG=618664,616716

Review-Url: https://codereview.chromium.org/2051953002
Cr-Original-Commit-Position: refs/heads/master@{#399130}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 06fdd11aadb98dd6e25c6aaace831998b23eae86