Implement Suborigin's 'unsafe-cookies' opt-out
According to the spec draft
(https://w3c.github.io/webappsec-suborigins/), by default, a
Document in a Suborigin is cookie-averse. However, there is an optional
directive 'unsafe-cookies' that causes the Suborigin to share
document.cookie fully with the physical origin.

This implements that opt-out directive and is a follow-up to the
cookie-averse implementation in
https://codereview.chromium.org/2020513002/. It adds the parsing for
'unsafe-cookies' and a simple set of checks in Document to bypass the
cookie-averse behavior.

BUG=336894
R=mkwst@chromium.org

Review-Url: https://codereview.chromium.org/2049863002
Cr-Original-Commit-Position: refs/heads/master@{#399566}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 15083af1018185cac2891d25cec7678cfabe676b