Add a method to LayoutTestResults to get a list of tests with new baselines.
The purpose of this change is to allow rebaseline commands to get a list of tests that need new baselines from a particular run of run-webkit-tests.

I'm assuming now that the set of tests with unexpected results is the set of tests that we want to fetch the new baselines for, although I might be wrong. For example, maybe if the expectation is CRASH and actual is PASS, then that would be unexpected, but maybe we wouldn't want to rebaseline -- is that right?

BUG=474273

Review-Url: https://codereview.chromium.org/2029893003
Cr-Original-Commit-Position: refs/heads/master@{#399843}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9f65556df7f2fbb2b096d3dfa49f1aa2ca13ab8