Reland: WTF: Implement explicit RefPtr::operator bool.
This is the modern alternative to the "safe bool" idiom. Some changes at
call sites are required, primarily because returning a boolean does not
cause an explicit conversion.

This is a reland of https://codereview.chromium.org/1891473002/ with no changes
aside from a rebase. There should be no more performance impact after the fix
in bug 607208.

BUG=607208

Review-Url: https://codereview.chromium.org/2063773002
Cr-Original-Commit-Position: refs/heads/master@{#399979}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 70843029b30a2566640848f072312cde61954ed3