Add WebDragData to blink::WebView::dragtargetDrop.
Different from Windows/Linux, the drag and drop data on Android is not available
when the drag starts. It is only available when the drop is performed (finger
releases from touch screen).

During DragEnter, the meta data of the drop data will be sent to renderer
to construct a WebDragData that contains only this meta data and it will be used by
blink.

This CL adds a WebDragData parameter to blink WebView::dragTargetDrop. When the
dragged data is dropped, the full drop data will be sent to renderer and then to
blink.

This CL also changes the logic around drop data filtering and permisson granting.
Different platforms (Aura, Mac, BrowserPlugin) will filter the drop data before
DragEnter, but not Android, because it is not necessary. The filtered drop data
will be cached and used at Drop for these platforms except Android.

Android logic for drop data will be in a follow up CL.

BUG=584789

Review-Url: https://codereview.chromium.org/1723763002
Cr-Original-Commit-Position: refs/heads/master@{#399989}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8915b509e11743c64a473db748181c62ded64258