Remove WorkerBackingThread::scriptCount
This CL removes scriptCount from WorkerBackingThread. The count was
needed for compositor workers, but we don't need it any more because
of compositor workers' design change.

Before this change, compositor workers are exempted from
TerminateExecution call because multiple workers can share the same
backing thread, but with this change TerminateExecution can be
called for a compositor worker when it doesn't start the shutdown
sequence in a timely manner.

BUG=487050

Review-Url: https://codereview.chromium.org/2053273002
Cr-Original-Commit-Position: refs/heads/master@{#400102}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 17686215e73797de165dd749f8f77d1388f1ede6